Extend gitignore to cover additional generated files. (#5)
The LaTeX compilation process generates intermediate files (e.g. with
extensions `.gtex` or `.aux`) as well as log files (with extension `.log`).

These files are not of any use after generating a PDF, so this commit clears
them out and extends the `.gitignore` to avoid committing them in the future.